# docs/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QUEST API",
    "version": "0.0.0-dev",
    "description": "Generated from the zod contracts in @quest/types (ADR-012). Conventions: docs/api/API_CONVENTIONS.md. Regenerate with `pnpm --filter @quest/api openapi:generate`."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "local"
    }
  ],
...
  "paths": {
...
    "/v1/me": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "The signed-in account",
        "operationId": "get-me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "email": {
                      "type": "string"
                    },
                    "emailVerified": {
                      "type": "boolean"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "PENDING_VERIFICATION",
                        "ACTIVE",
                        "DEACTIVATED",
                        "SUSPENDED",
                        "DELETION_REQUESTED",
                        "DELETED"
                      ]
                    },
                    "ageBand": {
                      "type": "string",
                      "enum": [
                        "TEEN_13_15",
                        "TEEN_16_17",
                        "ADULT"
                      ]
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "USER",
                          "SUPER_ADMIN",
                          "TRUST_SAFETY_LEAD",
                          "MODERATOR",
                          "SUPPORT",
                          "ANALYST",
                          "READ_ONLY"
                        ]
                      }
                    },
                    "onboarding": {
                      "type": "object",
                      "properties": {
                        "completed": {
                          "type": "boolean"
                        },
                        "nextStep": {
                          "type": "string",
                          "enum": [
                            "VERIFY_EMAIL",
                            "PROFILE",
                            "INTERESTS",
                            "DONE"
                          ]
                        }
                      },
                      "required": [
                        "completed",
                        "nextStep"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "deletionScheduledFor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "accountId",
                    "email",
                    "emailVerified",
                    "state",
                    "ageBand",
                    "roles",
                    "onboarding",
                    "createdAt",
                    "deletionScheduledFor"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION",
          "DELETION_REQUESTED"
        ]
      }
    },
...
    "/v1/me/consents": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Current consent state",
        "operationId": "get-me-consents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "current": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string",
                        "enum": [
                          "TERMS_OF_SERVICE",
                          "PRIVACY_POLICY",
                          "AGE_ATTESTATION",
                          "ANALYTICS",
                          "PERSONALISATION",
                          "MARKETING"
                        ]
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "granted": {
                            "type": "boolean"
                          },
                          "documentVersion": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "recordedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          }
                        },
                        "required": [
                          "granted",
                          "documentVersion",
                          "recordedAt"
                        ]
                      }
                    },
                    "requiresReprompt": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "TERMS_OF_SERVICE",
                          "PRIVACY_POLICY",
                          "AGE_ATTESTATION",
                          "ANALYTICS",
                          "PERSONALISATION",
                          "MARKETING"
                        ]
                      }
                    }
                  },
                  "required": [
                    "current",
                    "requiresReprompt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_ACCOUNT"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION",
          "DELETION_REQUESTED"
        ]
      },
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Record a consent decision",
        "operationId": "post-me-consents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "current": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string",
                        "enum": [
                          "TERMS_OF_SERVICE",
                          "PRIVACY_POLICY",
                          "AGE_ATTESTATION",
                          "ANALYTICS",
                          "PERSONALISATION",
                          "MARKETING"
                        ]
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "granted": {
                            "type": "boolean"
                          },
                          "documentVersion": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "recordedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          }
                        },
                        "required": [
                          "granted",
                          "documentVersion",
                          "recordedAt"
                        ]
                      }
                    },
                    "requiresReprompt": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "TERMS_OF_SERVICE",
                          "PRIVACY_POLICY",
                          "AGE_ATTESTATION",
                          "ANALYTICS",
                          "PERSONALISATION",
                          "MARKETING"
                        ]
                      }
                    }
                  },
                  "required": [
                    "current",
                    "requiresReprompt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_ACCOUNT"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "TERMS_OF_SERVICE",
                      "PRIVACY_POLICY",
                      "AGE_ATTESTATION",
                      "ANALYTICS",
                      "PERSONALISATION",
                      "MARKETING"
                    ]
                  },
                  "granted": {
                    "type": "boolean"
                  },
                  "documentVersion": {
                    "type": "string",
                    "pattern": "^[0-9]{4}-[0-9]{2}(\\.[0-9]+)?$"
                  }
                },
                "required": [
                  "type",
                  "granted"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/me/consents/history": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Consent history (append-only)",
        "operationId": "get-me-consents-history",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "consentId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "TERMS_OF_SERVICE",
                              "PRIVACY_POLICY",
                              "AGE_ATTESTATION",
                              "ANALYTICS",
                              "PERSONALISATION",
                              "MARKETING"
                            ]
                          },
                          "documentVersion": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^[0-9]{4}-[0-9]{2}(\\.[0-9]+)?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "granted": {
                            "type": "boolean"
                          },
                          "source": {
                            "type": "string",
                            "enum": [
                              "REGISTRATION",
                              "SETTINGS",
                              "REPROMPT",
                              "SUPPORT"
                            ]
                          },
                          "recordedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          }
                        },
                        "required": [
                          "consentId",
                          "type",
                          "documentVersion",
                          "granted",
                          "source",
                          "recordedAt"
                        ]
                      }
                    },
                    "pageInfo": {
                      "type": "object",
                      "properties": {
                        "nextCursor": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "nextCursor",
                        "hasMore"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pageInfo"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_ACCOUNT"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION",
          "DELETION_REQUESTED"
        ]
      }
    },
...
    "/v1/me/data-export/{exportId}": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Data export by id",
        "operationId": "get-me-data-export-by-exportId",
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exportId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "REQUESTED",
                        "PROCESSING",
                        "READY",
                        "EXPIRED",
                        "FAILED"
                      ]
                    },
                    "requestedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "completedAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "expiresAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "downloadUrl": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uri"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "exportId",
                    "status",
                    "requestedAt",
                    "completedAt",
                    "expiresAt",
                    "downloadUrl"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "REQUEST_OWN_DATA_EXPORT"
        ],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ]
      }
    },
    "/v1/me/deactivate": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Deactivate the account (sign-in reactivates)",
        "operationId": "post-me-deactivate",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [
          "MANAGE_OWN_ACCOUNT"
        ],
        "x-quest-account-states": [
          "ACTIVE"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}